Move test, add one new
